Autodelete sessions in SqlAlchemySessionInterface
The expired sessions are not deleted from the database in every case, so
the autodeletion of expired sessions has been implemented.

Deletion of the expired sessions (e.g. with a cronjob) would be feasible
in the application, but it is nicer to implement it directly in
flask-session.

The autodeletion be enabled by the `SESSION_AUTODELETE` config variable,
which defaults to `False`.

Major drawback: it slows down `open_session`.